mlx-app-converter/manual-tests/streamlit-vlm-fixture/pyproject.toml
[project]
name = "streamlit-vlm-fixture"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "streamlit",
    "transformers",
    "torch",
    "pillow",
]

[dependency-groups]
dev = [
    "pytest",
    "ruff",
    "ty",
]
